Adding more info around using strategies
I realized that the docs were pretty light on using strategies
outside of bitgatt, I wanted to make it clear how to do it so that
we could start getting bugs and fixing the interfaces.  In the future
it might make sense for bitgatt to provide some sort of a transaction
consumer interface that had hooks for strategies at defined points
before and after the transaction, but this will require additional
thinking.