Fixes # Crash with more than 1000 receivers
If a developer registered the receiver programmatically and in the
manifest this could lead to a crash on some phones where the intent
based scan seems to return results on different threads leading
to a start race that got through the atomic boolean get.

Added tests for this behavior, guard logging so that a dev knows that
this is going on, and made sure that this receiver could only be
registered once.

Tested in gucci gatt manually on a Pixel 3 XL as well as automated
testing

Interesting side-note, on the Pixel and well behaved phones
too many registrations of the same instance does not result
in a crash, only a log statement indicating that this instance
is already registered, however on some poorly behaved phones like
the Samsung A8 this results in a crash, this may be fixed in later
Android versions but at least some versions crash.