Fixes # [Internal]
We have discovered that at times the gatt db becomes corrupted,
requiring
the user to clear the "bluetooth" or the "Bluetooth Share" application
to prevent ANRs we have moved the gatt server start to a background
thread and will callback success or failure on startup, if your startup
fails, then you should ask your user to clear bluetooth share on their
phone.

Sadly, there is no programmatic way to do this.