On some devices, the gatt server behavior is somewhat unpredictable
In order to ensure consistent behavior we need to remove services
when bluetooth is turning off.  In addition, there was a bug in that
if you created a number of CompositeTransactions you could end up
holding onto a large number of active threads, it's better to only
start the threads when you will actually use them, and to ensure
that they are cancelled.