Busy main looper can ANR
In testing and in production we have seen ANRs that do not look like
ANRs in the log where the main looper is waiting on a lock and the
background thread that is posting to the main looper is also waiting
on the lock.  This leads to a deadlock situation where the main looper's
message queue is waiting on the synchronization barrier while the
message
queue is servicing the next linked work chain ... this chain can get
long while the main looper is waiting on kpoll and then can get stuck
in the while loop for more than the ANR timeout.

The simplest way to fix this is to move the callbacks for events that
occur asynchronously in the gatt client over to a new handler thread and
therefore a new MessageQueue that is far less busy.  This still involves
some overhead, so we will eventually stop using the handlerthread
entirely
and move the implementation to the linked blocking queue as we have
for the outbound events