PI scans must be stopped
The pending intent scanner creates a random UUID in the place
of app id since it isn't really our app doing the scanning, this
app id is placed in a scanner map with the pending intent context

It looks like if our app dies and comes back up we could end up wasting
gatt IFs because of how the code is written, so we really must clean up
the existing pending intent scan when we start a new one. We might be
able to do that by cancelling the old one with a newly created PI that
matches the app package and class or the old one.