Fix Android Stack Crash on Android 7
We thought originally that this crash was caused because
someone could clear services while we were adding
services.  This may have been the case, but we are still
observing crashes on Andorid 7, as this seems to coincide with
other odd behavior, we are assuming that the bluetooth service has
crashed leading to the internal NPE.

If this is the case it is unlikely that we will be able to handle
these failures by business logic alone.